chore: refactor test strcuture
