+++ frontend/appflowy_flutter/integration_test/desktop/cloud/cloud_runner.dart
@@ -1,28 +1,18 @@
-import 'document/document_delete_block_test.dart' as document_delete_block_test;
-import 'document/document_drag_block_test.dart' as document_drag_block_test;
-import 'document/document_option_actions_test.dart'
-    as document_option_actions_test;
+import 'document/document_test_runner.dart' as document_test_runner;
 import 'sidebar/sidebar_move_page_test.dart' as sidebar_move_page_test;
-import 'uncategorized/anon_user_continue_test.dart' as anon_user_continue_test;
-import 'uncategorized/appflowy_cloud_auth_test.dart'
-    as appflowy_cloud_auth_test;
-import 'uncategorized/empty_test.dart' as preset_af_cloud_env_test;
-import 'uncategorized/user_setting_sync_test.dart' as user_sync_test;
+import 'uncategorized/uncategorized_test_runner.dart'
+    as uncategorized_test_runner;
 import 'workspace/workspace_test_runner.dart' as workspace_test_runner;
 
 Future<void> main() async {
-  preset_af_cloud_env_test.main();
-  appflowy_cloud_auth_test.main();
-  user_sync_test.main();
-  anon_user_continue_test.main();
+  // uncategorized
+  uncategorized_test_runner.main();
 
   // workspace
-  workspace_test_runner.startTesting();
+  workspace_test_runner.main();
 
   // document
-  document_option_actions_test.main();
-  document_drag_block_test.main();
-  document_delete_block_test.main();
+  document_test_runner.main();
 
   // sidebar
   sidebar_move_page_test.main();
