+++ frontend/appflowy_flutter/integration_test/desktop/cloud/document/document_option_actions_test.dart
@@ -1,5 +1,7 @@
 import 'package:appflowy/env/cloud_env.dart';
 import 'package:appflowy/generated/locale_keys.g.dart';
+import 'package:appflowy/plugins/document/presentation/editor_plugins/actions/drag_to_reorder/draggable_option_button.dart';
+import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -96,6 +98,49 @@ void main() {
           r'^https:\/\/appflowy\.com\/app\/[a-f0-9-]{36}\/[a-f0-9-]{36}\?blockId=[A-Za-z0-9_-]+$',
         ),
       );
+    });
+
+    testWidgets('hover on the block and delete it', (tester) async {
+      await tester.initializeAppFlowy(
+        cloudType: AuthenticatorType.appflowyCloudSelfHost,
+      );
+      await tester.tapGoogleLoginInButton();
+      await tester.expectToSeeHomePageWithGetStartedPage();
+
+      // open getting started page
+      await tester.openPage(Constants.gettingStartedPageName);
+
+      // before delete
+      final path = [1];
+      final beforeDeletedBlock = tester.editor.getNodeAtPath(path);
+
+      // hover on the block and delete it
+      final optionButton = find.byWidgetPredicate(
+        (widget) =>
+            widget is DraggableOptionButton &&
+            widget.blockComponentContext.node.path.equals(path),
+      );
+
+      await tester.hoverOnWidget(
+        optionButton,
+        onHover: () async {
+          // click the delete button
+          await tester.tapButton(optionButton);
+        },
+      );
+      await tester.pumpAndSettle(Durations.short1);
+
+      // click the delete button
+      final deleteButton =
+          find.findTextInFlowyText(LocaleKeys.button_delete.tr());
+      await tester.tapButton(deleteButton);
+
+      // wait for the deletion
+      await tester.pumpAndSettle(Durations.short1);
+
+      // check if the block is deleted
+      final afterDeletedBlock = tester.editor.getNodeAtPath([1]);
+      expect(afterDeletedBlock.id, isNot(equals(beforeDeletedBlock.id)));
     });
   });
 }
