+++ frontend/appflowy_flutter/integration_test/desktop/cloud/document/document_test_runner.dart
@@ -1,0 +1,9 @@
+import 'package:integration_test/integration_test.dart';
+
+import 'document_option_actions_test.dart' as document_option_actions_test;
+
+void main() {
+  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+
+  document_option_actions_test.main();
+}
